Remove unused code and split LogContent
