+++ studio/src/components/Timeline/DetailsList/TimelineDetailsList/TimelineListDetails.tsx
@@ -26,7 +26,7 @@ function TimelineListDetailsComponent({
   const tree = useMemo(() => convertToTree(waterfall), [waterfall]);
 
   return (
-    <div className="grid gap-4 min-h-0 mt-0">
+    <div className="grid gap-1 min-h-0 mt-0">
       <div className="flex gap-12 border-b pb-2">
         <label className="flex items-center space-x-2 cursor-pointer">
           <span className="text-xs text-muted-foreground ">Tree view</span>
