+++ studio/src/components/ui/pagination.tsx
@@ -5,8 +5,7 @@ import {
 } from "@radix-ui/react-icons";
 import * as React from "react";
 
-import type { ButtonProps } from "@/components/ui/button";
-import { buttonVariants } from "@/components/ui/button/variants";
+import { type ButtonProps, buttonVariants } from "@/components/ui/button";
 import { cn } from "@/utils";
 
 const Pagination = ({ className, ...props }: React.ComponentProps<"nav">) => (
